+++ niryo_robot_hardware_stack/ttl_driver/include/ttl_driver/mock_dxl_driver.hpp
@@ -73,14 +73,14 @@ class MockDxlDriver : public AbstractDxlDriver
         // ram read
         virtual int readPosition(uint8_t id, uint32_t &present_position) override;
         virtual int readTemperature(uint8_t id, uint32_t &temperature) override;
-        virtual int readVoltage(uint8_t id, uint32_t &voltage) override;
+        virtual int readVoltage(uint8_t id, double &voltage) override;
         virtual int readHwErrorStatus(uint8_t id, uint32_t &hardware_status) override;
 
         virtual int syncReadPosition(const std::vector<uint8_t> &id_list, std::vector<uint32_t> &position_list) override;
 
         virtual int syncReadFirmwareVersion(const std::vector<uint8_t> &id_list, std::vector<std::string> &firmware_list) override;
         virtual int syncReadTemperature(const std::vector<uint8_t> &id_list, std::vector<uint32_t> &temperature_list) override;
-        virtual int syncReadVoltage(const std::vector<uint8_t> &id_list, std::vector<uint32_t> &voltage_list) override;
+        virtual int syncReadVoltage(const std::vector<uint8_t> &id_list, std::vector<double> &voltage_list) override;
         virtual int syncReadHwErrorStatus(const std::vector<uint8_t> &id_list, std::vector<uint32_t> &hw_error_list) override;
 
         // AbstractDxlDriver interface
@@ -108,7 +108,7 @@ class MockDxlDriver : public AbstractDxlDriver
         {
           uint32_t       position{};
           uint32_t       temperature{};
-          uint32_t       voltage{};
+          double         voltage{};
           uint32_t       min_position{};
           uint32_t       max_position{};
           uint16_t       model_number{};
@@ -298,7 +298,7 @@ int MockDxlDriver::readTemperature(uint8_t id, uint32_t& temperature)
     return COMM_SUCCESS;
 }
 
-int MockDxlDriver::readVoltage(uint8_t id, uint32_t& voltage)
+int MockDxlDriver::readVoltage(uint8_t id, double& voltage)
 {
     voltage = _map_fake_registers.at(id).voltage;
     return COMM_SUCCESS;
@@ -349,7 +349,7 @@ int MockDxlDriver::syncReadTemperature(const std::vector<uint8_t> &id_list, std:
     return COMM_SUCCESS;
 }
 
-int MockDxlDriver::syncReadVoltage(const std::vector<uint8_t> &id_list, std::vector<uint32_t> &voltage_list)
+int MockDxlDriver::syncReadVoltage(const std::vector<uint8_t> &id_list, std::vector<double> &voltage_list)
 {
     std::map<uint8_t, uint8_t> countMap;
     for (auto & id : id_list)
